Explore Arduino limitations
Control for the Python script by replacing it with a simple bash script. The Arduino bugs out when we sleep less than 30 ms. This isn't so bad, but I'd like to try to get it back down to being fine with one command per millisecond; that way for every image we'll have a command that's close to it in time. I'm not sure if the Arduino ever did handle 1ms per command, but I think it was able to; the Arduino code was a lot simpler then; although it didn't check for bad inputs, which allowed the car to do dangerous things.