Explore Arduino Limitations (2)
I commented out unneeded Arduino instructions in hope of reducing its load. I found that after doing so the bash script works with the Arduino as long as the bash script has a sleep instruction. Without a sleep instruction, data is lost and the Arduino ends up with invalid commands. What's strange is that an arbitarily small sleep instruction prevents the data loss.